+++ src/count-modals/number-modal.svelte
@@ -2,7 +2,7 @@
     import { createEventDispatcher } from 'svelte';
     import Textfield from '@smui/textfield';
     import Button, { Label } from '@smui/button';
-    import Modal from '../util/modal.svelte';
+    import Modal from './modal.svelte';
 
     export let count;
 
@@ -12,12 +12,8 @@
 <Modal>
     <span slot="name"><slot></slot></span>
 
-    <h3>Choose Start Date</h3>
-    <Textfield bind:value={count.startDate} Label="Start Date" type="date"/>
-    <br/><br/>
-
-    <h3>Choose End Date</h3>
-    <Textfield bind:value={count.endDate} Label="End Date" type="date"/>
+    <h3>Choose Value</h3>
+    <Textfield bind:value={count.value} Label="Value"/>
 
     <div slot="actions">
         <Button on:click={dispatch('submit', {count})} variant="unelevated"><Label>Submit</Label></Button>
