--- src/shared/selected-tab.store.js
@@ -1,12 +1,0 @@
-import { writable } from 'svelte/store';
-
-function createSelectedTab() {
-    const { subscribe, set } = writable(undefined);
-
-    return {
-        subscribe,
-        set: (selectedTab) => set(selectedTab)
-    }
-}
-
-export const selectedTab = createSelectedTab();
