+++ src/stores/selected-tab.store.js
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const selectedTab = writable(undefined);
