+++ src/stores/status.store.js
